### 2025-10-05 #9143 @editorialbot
[PRE REVIEW]: GlobalSearch-rs: A global optimization library written in Rust
**Submitting author:** @GermanHeim (<a href="http://orcid.org/0009-0007-9834-7924">Germán Martín Heim</a>)
**Repository:** https://github.com/GermanHeim/globalsearch-rs
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.3.0
