### 2025-10-07 #9143 @GermanHeim
Hello @nkrusch 👋,
Thank you for agreeing to edit!

Regarding reviewers, I can suggest the following:

- Rohit Goswami (HaoZeke)
- Mathieu Besançon (matbesancon)
- Leonardo Giannantoni (lgiannantoni)
- Camilo Chacón Sartori (camilochs)

I could also suggest Rémi Lafage (relf), who has previously published in JOSS and is the author of one of the crates I use. I didn’t see him listed as a reviewer, but he might be open to participating.

Thanks again,
Germán
