### 2025-10-08 #9143 @nkrusch
Hello @matbesancon, @lgiannantoni, @camilochs and @relf:

would you be interested in reviewing this submission for JOSS?
