### 2025-10-13 GermanHeim/globalsearch-rs#8 @jonaspleyer
[feature] support for arbitrary floating point types (eg. f32, f64)
There exist many use-cases for `f32` computations such as machine learning and training of neural networks. For a more seamless integration with such users, `globalsearch` could opt to support arbitrary floating point types (eg. `f32` and `f64`). I think that this idea is worthwhile but I am aware of the implications and effort that has to be taken to achieve this.

**Describe the solution you'd like**
I would be in favor of using generics

```rust
fn delta<F>(mut self, delta: F) -> Self {...}
```

Another option would be to define a global type and feature-guard it.
```rust
#[cfg(feature = "f32")]
pub type Float = f32;
#[cfg(feature = "f64")]
pub type Float = f64;

#[cfg(all(feature = "f32", feature = "f64")))]
compile_error!("f32 and f64 feature are mutually exclusive!");

fn delta(mut self, delta: Float) -> Self {...}
```
I do not prefer this solution as it would be challenging for mixed-case usages.
Furthermore, packages that want to build on `globalsearch` now do not have the option to use generics anymore.

**Additional context**

https://github.com/openjournals/joss-reviews/issues/9234
