### 2025-10-15 c3e82b0
Clarify limitations of GlobalSearch-rs
Added explanation that GlobalSearch-rs cannot guarantee finding the true global minimum
