+++ paper/paper.md
@@ -49,7 +49,9 @@ Rust’s ownership model and concurrency features enable safe and efficient exec
 
 # Library features
 
-`GlobalSearch-rs` implements scatter search metaheuristics, which balances diversification and intensification strategies to navigate complex objective landscapes. By leveraging these techniques, the algorithm improves the exploration of global minima while refining promising candidate solutions through local search. The library integrates with both the `cobyla` and `argmin` crates. These Rust-based optimization libraries provide a range of local NLP solvers. Supported methods include COBYLA, LBFGS, Nelder-Mead, Steepest Descent, Newton-Conjugate-Gradient and Trust Region solvers, ensuring flexibility in handling different types of objective functions [@drones7080487].
+`GlobalSearch-rs` implements scatter search metaheuristics, which balances diversification and intensification strategies to navigate complex objective landscapes. By leveraging these techniques, the algorithm improves the exploration of global minima while refining promising candidate solutions through local search. However, as with other stochastic global optimization algorithms, `GlobalSearch-rs` cannot guarantee that the best solution found is the true global minimum. Instead, it aims to identify high-quality solutions with high probability. Users can increase confidence by performing multiple independent runs or using different random seeds.
+
+The library integrates with both the `cobyla` and `argmin` crates. These Rust-based optimization libraries provide a range of local NLP solvers. Supported methods include COBYLA, LBFGS, Nelder-Mead, Steepest Descent, Newton-Conjugate-Gradient and Trust Region solvers, ensuring flexibility in handling different types of objective functions [@drones7080487].
 
 Parallel execution is a core feature of `GlobalSearch-rs`, utilizing the `Rayon` library to distribute computation across multiple threads. This has the potential to enhance computational efficiency, enabling users to tackle large-scale optimization problems with improved performance compared to single-threaded alternatives. To further support long-running tasks, the library provides checkpointing capabilities, allowing users to save and resume optimization states without restarting from scratch.
 
