### 2025-10-18 GermanHeim/globalsearch-rs#13 @relf
[docs] Multistart terminology is not used
### 📖 Documentation Issue
My main comment would be to mention from the start that globalsearch pertain to
multistart optimization methods as it is a well-known strategy to mitigate local optimization
(I did not know about the OQNLP algorithm and was a bit puzzled by "scatter search metaheuristics" and "NLP solver" vocabulary to convey the idea of a restarted local optimizations algorithm).

### 📝 Suggested Improvement
I've search "multistart" through the project and the paper and found the only occurences are in the main reference
"Scatter Search and Local NLP Solvers: A Multistart Framework for Global Optimization".
Why not use "multistart" in the title of the paper (GlobalSearch-rs: a multistart framework for
global optimization written in rust?) ? At least, it should be mentioned in the first lines of the summary even before OQNLP. Something like "globalsearch implements a multistart optimization strategy. It is a port in Rust of the OptQuest/NLP algorithm, combining..."

### 🔗 Related Links
I think using "multistart" terminology will help mitigate the misunderstanding between global vs local mentioned by @matbesancon in his [comment](https://github.com/openjournals/joss-reviews/issues/9234#issuecomment-3400504919)

https://github.com/openjournals/joss-reviews/issues/9234
