### 2025-10-18 f2898e8
Clarify multistart framework
