### 2025-10-18 GermanHeim/globalsearch-rs#13 @GermanHeim
Hello!

Thank you for the feedback, I agree that the terminology could be more precise. I’ve made some improvements to address this in commit https://github.com/GermanHeim/globalsearch-rs/commit/c3e82b066b7439b3c2842497774ffa4feffb138d and pull request https://github.com/GermanHeim/globalsearch-rs-website/pull/1

A more detailed explanation of how the algorithm works is available in the [Website's overview](https://germanheim.github.io/globalsearch-rs-website/getting-started/overview/), which is actually in another [repository](https://github.com/GermanHeim/globalsearch-rs-website). This could explain why you didn't find much information when searching the main repo.

I also applied your suggestion about introducing the "multistart" terminology earlier in the paper. You can see these changes in commit https://github.com/GermanHeim/globalsearch-rs/commit/f2898e8bf92aa549d6cc6b69a1e4353372587f60. Does this make the terminology more straightforward?
