### 2025-10-18 #9234 @GermanHeim
👋 @nkrusch - We've changed the title of the paper. Should we manually edit the submission title or is that done automatically?
