### 2025-10-19 GermanHeim/globalsearch-rs#14 @amcandio
[JOSS review] Manuscript feedback
Hi! Here is my feedback for the manuscript as part of JOSS paper review (https://github.com/openjournals/joss-reviews/issues/9234)

### Missing references

It would be great to add references to dependencies argmin (https://argmin-rs.org/book/) and cobyla (Powell 1994). I see scipy cites a reference implementation of cobyla https://docs.scipy.org/doc/scipy/reference/optimize.minimize-cobyla.html. You can use the reference you consider the most adequate.

There is other software mentioned that should be cited suchas numpy and jax.

### Other tools

NLopt and pagmo2 are two popular libraries with both rust and python bindings. I think they should be referenced in the manuscript. From my understanding, those don't provide an implementation for "OQNLP" and it would help strenghten the statement of need.

### Writing

Not much to comment here, manuscript seems pretty well-written. I do see adjetives such as "modern" or "accessible" which are not measurable and can be percieved as subjective for a scientific paper. For words like "extensible", manuscript should explain why library is extensible.

Regarding sentence "The algorithm then enters the stage two...". This is a bit odd because there is no previous mention to stages. You can update the initial parragraph explaining the algorithm has two stages. In any case, this section could be consider out of scope for JOSS

### Scope

From [JOSS guidelines](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain):
> Given this format, a “full length” paper is not permitted, and software documentation such as API (Application Programming Interface) functionality should not be in the paper and instead should be outlined in the software documentation.

I believe the section "Example: Optimizing the Six-Hump Camel Function" is valuable for library's documentation but it is too deep into the details for JOSS.
