### 2025-10-19 GermanHeim/globalsearch-rs#15 @amcandio
[JOSS review] Python example is outdated
As part of https://github.com/openjournals/joss-reviews/issues/9234 review, I'm getting this error when following python example https://germanheim.github.io/globalsearch-rs-website/python/running-an-optimization-problem/

```
... problem = gs.Problem(objective=objective_function, variable_bounds=variable_bounds)
...
Traceback (most recent call last):
  File "<python-input-0>", line 12, in <module>
    problem = gs.Problem(objective=objective_function, variable_bounds=variable_bounds)
              ^^^^^^^^^^
AttributeError: module 'pyglobalsearch' has no attribute 'Problem'. Did you mean: 'PyProblem'?
```

Regarding the solution example, you don't need to change this but it might be good to either pick one with a non-zero solution or either clarify in the doc what is the expected output.

It might just be me, but the moment I see:
```
Solution Set
━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
Total solutions: 1
Best objective value: 0.00000000e0
━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
Solution #1
  Objective: 0.00000000e0
  Parameters: [0.0, 0.0]
```

the first thing I think is that there is a bug. Then I checked what the objective function was and it was fine :)
