### 2025-10-19 41e7f8b
Add citations for NumPy, JAX, argmin, and cobyla
Added BibTeX entries for NumPy, JAX, argmin, and cobyla to paper.bib and updated paper.md to cite these libraries
