+++ paper/paper.bib
@@ -71,3 +71,66 @@
   volume         = {7},
   year           = {2023}
 }
+
+@article{numpy,
+  title  = {Array programming with {NumPy}},
+  author = {Charles R. Harris and K. Jarrod Millman and St{\'{e}}fan J.
+                van der Walt and Ralf Gommers and Pauli Virtanen and David
+                Cournapeau and Eric Wieser and Julian Taylor and Sebastian
+                Berg and Nathaniel J. Smith and Robert Kern and Matti Picus
+                and Stephan Hoyer and Marten H. van Kerkwijk and Matthew
+                Brett and Allan Haldane and Jaime Fern{\'{a}}ndez del
+                R{\'{i}}o and Mark Wiebe and Pearu Peterson and Pierre
+                G{\'{e}}rard-Marchant and Kevin Sheppard and Tyler Reddy and
+                Warren Weckesser and Hameer Abbasi and Christoph Gohlke and
+                Travis E. Oliphant},
+  year          = {2020},
+  month         = sep,
+  journal       = {Nature},
+  volume        = {585},
+  number        = {7825},
+  pages         = {357--362},
+  doi           = {10.1038/s41586-020-2649-2},
+  publisher     = {Springer Science and Business Media {LLC}},
+  url           = {https://doi.org/10.1038/s41586-020-2649-2}
+}
+
+@misc{jax,
+  author    = {James Bradbury and Roy Frostig and Peter Hawkins and Matthew James Johnson and Chris Leary and Dougal Maclaurin and George Necula and Adam Paszke and Jake Vander{P}las and Skye Wanderman-{M}ilne and Qiao Zhang},
+  journal   = {GitHub repository},
+  title     = {{JAX}: composable transformations of {P}ython+{N}um{P}y programs},
+  publisher = {GitHub},
+  url       = {http://github.com/jax-ml/jax},
+  version   = {0.3.13},
+  year      = {2018}
+}
+
+@misc{jax,
+  author    = {James Bradbury and Roy Frostig and Peter Hawkins and Matthew James Johnson and Chris Leary and Dougal Maclaurin and George Necula and Adam Paszke and Jake Vander{P}las and Skye Wanderman-{M}ilne and Qiao Zhang},
+  journal   = {GitHub repository},
+  publisher = {GitHub},
+  title     = {{JAX}: composable transformations of {P}ython+{N}um{P}y programs},
+  url       = {http://github.com/jax-ml/jax},
+  version   = {0.3.13},
+  year      = {2018}
+}
+
+@misc{argmin,
+  author    = {Stefan Kroboth},
+  journal   = {GitHub repository},
+  publisher = {GitHub},
+  title     = {{argmin{}}: {M}athematical optimization in pure {R}ust.},
+  url       = {https://github.com/argmin-rs/argmin},
+  version   = {0.11.0},
+  year      = {2017}
+}
+
+@misc{cobyla,
+  author    = {Rémi Lafage},
+  journal   = {GitHub repository},
+  publisher = {GitHub},
+  title     = {{cobyla -} a pure {R}ust implementation.},
+  url       = {https://github.com/relf/cobyla/},
+  version   = {0.7.0},
+  year      = {2021}
+}
