### 2025-10-19 b3f1a3a
Clarify GlobalSearch-rs optimization process and stages
