### 2025-10-19 389db67
Add citations for NLopt and pagmo2 libraries
