### 2025-11-01 #9234 @jonaspleyer
Some comments concerning the paper:

1. You show two logos in separate Figures. Consider to combine them into one figure. It could also be possible to have a logo which combines the two showing "PyGlobalSearch" and "GlobalSearch-rs" alongside each other with the graphical visualization.
2. I consider the example "Optimizing the Six-Hump Camel Function" as an "original result" (as specified in the review checklist). I have found the examples within the Rust code and python code but could not see any visualizations such as shown in the paper. Please add the code to visualize the results.
3. The summary states (line 11): "[..] by leveraging [..] Rust’s memory safety guarantees." I do not see the clear connection of how the memory-safety of Rust contributes to the goals of this package. In contrast, the other mentioned aspect "[..] efficient parallel computing [..]" is better understandable.
4. The "Statement of need" states (line 27) "[..] no actively maintained open-source alternative currently exists." Clearly there are libraries such as `scipy.optimize` which are actively maintained and provide global optimization routines. This sentence should be more specific. I believe that you should target the particular algorithm which you implemented.
5. Line 36. either typo or imho awkward formulation: "[..] a suitability further supported by recent work demonstrating [..]"
6. Line 41. "[..] improves the exploration of global minima [..]" What is improved? Speed of convergence, accuracy, confidence, etc. Consider, linking to papers which explain the two terms "diversification" and "intensification".
7. Line 51: "This has the potential to enhance computational efficiency [..]" I believe that this statement is not entirely correct. Parallelization usually does not provide better efficiency but rather an overall speedup as stated just afterwards.
8. Line 70 consider using `$\approx$` instead of an equal sign. Furthermore, denote the two minima with distinct variable names for mathematical distinction. Suggestion: $x^*_\pm\approx(0.0989, \pm0.7126)$ `$x^*_\pm\approx(0.0989, \pm0.7126)$`

In general I find the paper well written. Most of the comments above are of minor nature and should improve the clarity of the paper. However at this point in time, the paper does not contain a "Related Software" or "State of Field" section which comments on existing solutions. I believe that this section should be added as stated  guidelines state as well.

I have furthermore dealt in detail with the code itself and was very pleased to work on said code base. All steps from installing, to running tests did work as expected.
