### 2025-11-01 #9234 @nkrusch
Adding to @jonaspleyer's previous comment, the discussion of related solutions should be in ["Statement of need"](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain):

> A Statement of need section that clearly illustrates the research purpose of the software and places it in the context of related work.

Briefly, there are philosophical differences between the JOSS paper format and papers that present theoretical results. The JOSS paper is meant to highlight a software contribution (with a limited word count). The convention is to present the difference to related works under the statement of need.

(_Meta comment:_ Each submission goes through a series of editorial checks before a submission is allowed to proceed to the review phase. This means that, beyond the discussion in the paper, there are complementary scope and quality checks that the submission must satisfy. As an editor, I have found that awareness of this procedure is important, so I wanted to mention it explicitly.)
