### 2025-11-01 033c908
Replace separate logo images with combined logos
