### 2025-11-01 f274dda
Update figure name
