### 2025-11-01 a087a46
Improve mathematical notation
Improve mathematical notation and minor fixes
