### 2025-11-01 8bd8d9d
Improve mathematical notation
