### 2025-11-02 be91fb5
Update paper with new reference and feature clarifications
Added a new reference for scatter search metaheuristics (Laguna2005) to the bibliography and cited it in the manuscript. Clarified the description of diversification and intensification strategies, improved explanations of Rust's parallelism, and specified the lack of open-source OQNLP implementations. Enhanced the discussion of library features and updated language for accuracy and clarity.

This was based on Jonas Pleyer's comment in https://github.com/openjournals/joss-reviews/issues/9234#issuecomment-3475333014.
