+++ paper/paper.bib
@@ -157,3 +157,20 @@
   version   = {2.10.0},
   year      = {2007}
 }
+
+@article{Laguna2005,
+  abstract = {Scatter search is an evolutionary method that, unlike genetic algorithms, operates on a small set of solutions and makes only limited use of randomization as a proxy for diversification when searching for a globally optimal solution. The scatter search framework is flexible, allowing the development of alternative implementations with varying degrees of sophistication. In this paper, we test the merit of several scatter search designs in the context of global optimization of multimodal functions. We compare these designs among themselves and choose one to compare against a well-known genetic algorithm that has been specifically developed for this class of problems. The testing is performed on a set of benchmark multimodal functions with known global minima.},
+  author   = {Laguna, Manuel and Mart{í}, Rafael},
+  day      = {01},
+  doi      = {10.1007/s10898-004-1936-z},
+  issn     = {1573-2916},
+  journal  = {Journal of Global Optimization},
+  month    = {Oct},
+  number   = {2},
+  pages    = {235-255},
+  title    = {Experimental Testing of Advanced Scatter Search Designs for Global Optimization of Multimodal Functions},
+  url      = {https://doi.org/10.1007/s10898-004-1936-z},
+  volume   = {33},
+  year     = {2005}
+}
+
