### 2025-11-02 #9234 @amcandio
No further comments from my side, this is great work! The only thing that is unclear to me is whether the example is out of scope for a JOSS paper.

Usually they focus on the functionality and explaining the algorithm step by step seems too deep into the details. What do you think @nkrusch ?
