### 2025-11-02 GermanHeim/globalsearch-rs#14 @amcandio
Thanks! I'm still not convinced with the example so let's discuss with the editor. I don't feel strongly about it so I'm fine accepting if the editor is okay with it.
