### 2025-11-03 #9234 @relf
I've finished my review. Just two minor comments on the paper which is otherwise well-written:
1. `pagmo2` does not seem to have a Rust API as stated in the paper (or I didn't find it)
2. The Figure 1 presenting the logos should be introduced and referenced in the text (or else removed)

Regarding the project, everything (code, documentation, examples, automated tests) is in place and adheres to high-quality standards. @GermanHeim, well done!
