### 2025-11-03 #9234 @nkrusch
@amcandio,

> Usually they focus on the functionality and explaining the algorithm step by step seems too deep into the details.

Generally, a JOSS paper aims to describe high-level functionality in non-specialist language, with more detailed information deferred to the software documentation. However, there is some flexibility around how to construct the paper. If the algorithm is stable and concise enough to explain in the paper (given the language and length requirements), I would say it is fine to include it. (I will do my detailed read after the reviewers, but based on a "light read", it seems fine for this paper).
