### 2025-11-03 48aa730
Implement feedback
Implement feedback from https://github.com/openjournals/joss-reviews/issues/9234#issuecomment-3479621055
