+++ paper/paper.bib
@@ -135,19 +135,6 @@
   year      = {2021}
 }
 
-@article{Biscani2020,
-  author    = {Francesco Biscani and Dario Izzo},
-  doi       = {10.21105/joss.02338},
-  journal   = {Journal of Open Source Software},
-  number    = {53},
-  pages     = {2338},
-  publisher = {The Open Journal},
-  title     = {A parallel global multiobjective framework for optimization: pagmo},
-  url       = {https://doi.org/10.21105/joss.02338},
-  volume    = {5},
-  year      = {2020}
-}
-
 @misc{nlopt,
   author    = {Steven G. Johnson},
   journal   = {GitHub repository},
