### 2025-11-05 #9234 @nkrusch
@GermanHeim, paper comments:

* In [Figure 1 caption](https://github.com/GermanHeim/globalsearch-rs/blame/paper/paper/paper.md#L52): the labels (left) and (right) are reversed.
* The noun form at [L45](https://github.com/GermanHeim/globalsearch-rs/blame/paper/paper/paper.md#L56C19-L56C91): _metaheuristics_ (plural noun) followed by _balances_ (singular verb); the form should match.
* [same sentence](https://github.com/GermanHeim/globalsearch-rs/blame/paper/paper/paper.md#L56): the parenthesized expression "(deeper search around promising areas using)" is confusing; i.e., word _using_ seems out of place.
* In references, L106, L119: use HTTPS in the address `http://github.com/jax-ml/jax` and `http://github.com/stevengj/nlopt`
