### 2025-11-05 e64af5f
Implement feedback
Implement the feedback from https://github.com/openjournals/joss-reviews/issues/9234#issuecomment-3491260120
