+++ paper/paper.bib
@@ -100,17 +100,7 @@
   journal   = {GitHub repository},
   title     = {{JAX}: composable transformations of {P}ython+{N}um{P}y programs},
   publisher = {GitHub},
-  url       = {http://github.com/jax-ml/jax},
-  version   = {0.3.13},
-  year      = {2018}
-}
-
-@misc{jax,
-  author    = {James Bradbury and Roy Frostig and Peter Hawkins and Matthew James Johnson and Chris Leary and Dougal Maclaurin and George Necula and Adam Paszke and Jake Vander{P}las and Skye Wanderman-{M}ilne and Qiao Zhang},
-  journal   = {GitHub repository},
-  publisher = {GitHub},
-  title     = {{JAX}: composable transformations of {P}ython+{N}um{P}y programs},
-  url       = {http://github.com/jax-ml/jax},
+  url       = {https://github.com/jax-ml/jax},
   version   = {0.3.13},
   year      = {2018}
 }
@@ -140,7 +130,7 @@
   journal   = {GitHub repository},
   publisher = {GitHub},
   title     = {The {NLopt} nonlinear-optimization package},
-  url       = {http://github.com/stevengj/nlopt},
+  url       = {https://github.com/stevengj/nlopt},
   version   = {2.10.0},
   year      = {2007}
 }
