### 2025-11-05 c8b93ec
GlobalSearch-rs v0.4.0
