+++ python/docs/conf.py
@@ -9,7 +9,7 @@ sys.path.insert(0, os.path.abspath(".."))
 project = "PyGlobalSearch"
 copyright = "2025, Germán Martín Heim"
 author = "Germán Martín Heim"
-release = "0.2.0"
+release = "0.4.0"
 
 # -- General configuration ---------------------------------------------------
 
