+++ README.md
@@ -54,7 +54,7 @@ Add this to your `Cargo.toml`:
 
 ```toml
 [dependencies]
-globalsearch = "0.3.0"
+globalsearch = "0.4.0"
 ```
 
 Or use `cargo add globalsearch` in your project directory.
