+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "globalsearch"
-version = "0.3.0"
+version = "0.4.0"
 edition = "2021"
 authors = ["Germán Martín Heim <german.martin.heim@gmail.com>"]
 license = "MIT"
