### 2025-11-05 #9234 @GermanHeim
Hello,
I just wanted to thank you all now that the paper is being published.

@jonaspleyer, @relf, @amcandio, your reviews really helped strengthen the library and pushed me to improve aspects I hadn’t considered. I can definitely say that the software presented is better thanks to you all! I also appreciate @nkrusch's guidance throughout the process, and I’m grateful to @matbesancon for the helpful discussion about the terminology used throughout the paper and library.

It was a genuinely good experience, and a pleasure to work with you!
