### 2025-11-05 71ddaa4
minor change in paper.md
