### 2025-11-05 #9234 @danielskatz
Congratulations to @GermanHeim (Germán Martín Heim) on your JOSS publication!!

And thanks very much to @jonaspleyer, @relf, and @amcandio for reviewing, and to @nkrusch for editing!
JOSS completely depends on volunteers; we couldn't do this without you
